Moves input and label components into app/components/forms
