+++ app/shops/new/shop-form.jsx
@@ -1,10 +1,10 @@
 'use client'
 import { useState } from "react";
 import { createClientComponentClient } from "@supabase/auth-helpers-nextjs";
-import Input from "./input";
+import Input from "../../components/forms/input";
 import styles from "./ShopForm.module.css";
 import { useRouter } from "next/navigation";
-import Label from "./label";
+import Label from "../../components/forms/label";
 // the form to create a new shop
 // Info needed: owner_id, name, description, items (just initialize the array)
